requirements.txt
wheel
flask
yfinance
TA-Lib
pandas
python-dotenv